+++ crates/forge/tests/e2e/running.rs
@@ -843,6 +843,7 @@ fn should_panic() {
     let snapbox = test_runner();
 
     let output = snapbox.current_dir(&temp).assert().code(1);
+
     assert_stdout_contains!(
         output,
         indoc! { r"
@@ -850,35 +851,53 @@ fn should_panic() {
         [..]Finished[..]
 
 
-        Collected 6 test(s) from should_panic_test package
+        Collected 8 test(s) from should_panic_test package
         Running 0 test(s) from src/
-        Running 6 test(s) from tests/
-        [PASS] tests::should_panic_test::should_panic_no_data [..]
-
-        Success data:
-            original value: [0], converted to a string: []
+        Running 8 test(s) from tests/
+        [FAIL] tests::should_panic_test::expected_panic_but_didnt
+        
+        Failure data:
+            Expected to panic but didn't
 
         [PASS] tests::should_panic_test::should_panic_check_data [..]
         [PASS] tests::should_panic_test::should_panic_multiple_messages [..]
+        [PASS] tests::should_panic_test::should_panic_no_data [..]
+        
+        Success data:
+            original value: [0], converted to a string: []
+        
         [FAIL] tests::should_panic_test::should_panic_with_non_matching_data
-
+        
         Failure data:
             Incorrect panic data
             Actual:    [8111420071579136082810415440747] (failing check)
             Expected:  [0] ()
-
+        
+        [FAIL] tests::should_panic_test::expected_panic_but_didnt_with_expected
+        
+        Failure data:
+            Expected to panic but didn't
+            Expected panic data:  [8903707727067478891290643490661] (panic message)
+        
+        [FAIL] tests::should_panic_test::expected_panic_but_didnt_with_expected_multiple
+        
+        Failure data:
+            Expected to panic but didn't
+            Expected panic data:  [8903707727067478891290643490661, 2340509922561928411394884117817189] (panic message, second message)
+        
         [FAIL] tests::should_panic_test::didnt_expect_panic
-
+        
         Failure data:
             original value: [156092886226808350968498952598218238307], converted to a string: [unexpected panic]
-
-        [FAIL] tests::should_panic_test::expected_panic_but_didnt
-        Tests: 3 passed, 3 failed, 0 skipped, 0 ignored, 0 filtered out
-
+        
+        Tests: 3 passed, 5 failed, 0 skipped, 0 ignored, 0 filtered out
+        
         Failures:
-            tests::should_panic_test::should_panic_with_non_matching_data
-            tests::should_panic_test::didnt_expect_panic
             tests::should_panic_test::expected_panic_but_didnt
+            tests::should_panic_test::should_panic_with_non_matching_data
+            tests::should_panic_test::expected_panic_but_didnt_with_expected
+            tests::should_panic_test::expected_panic_but_didnt_with_expected_multiple
+            tests::should_panic_test::didnt_expect_panic
         "}
     );
 }
@@ -935,7 +954,7 @@ fn incompatible_snforge_std_version_warning() {
     assert_stdout_contains!(
         output,
         indoc! {r"
-        [WARNING] Package snforge_std version does not meet the recommended version requirement =0.13.1, [..]
+        [WARNING] Package snforge_std version does not meet the recommended version requirement =0.14.0, [..]
         [..]Compiling[..]
         [..]Finished[..]