+++ crates/forge/tests/integration/mod.rs
@@ -22,6 +22,7 @@ mod should_panic;
 mod signing;
 mod spoof;
 mod spy_events;
+mod store_load;
 mod syscalls;
 mod test_state;
 mod trace;