+++ crates/sncast/src/response/print.rs
@@ -1,4 +1,5 @@
-use anyhow::Result;
+use anyhow::{Error, Result};
+use console::style;
 use serde_json::Value;
 use starknet::core::types::FieldElement;
 use std::{collections::HashMap, fmt::Display, str::FromStr};
@@ -161,6 +162,11 @@ fn apply_numbers_formatting(value: OutputValue, formatting: NumbersFormat) -> Ou
             OutputValue::Array(formatted_arr)
         }
     }
+}
+
+pub fn print_as_warning(error: &Error) {
+    let warning_tag = style("Warning:").color256(11);
+    println!("{warning_tag} {error}");
 }
 
 #[cfg(test)]