+++ CHANGELOG.md
@@ -11,6 +11,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
+- `store` and `load` cheatcodes
+- dependency on [universal-sierra-compiler](https://github.com/software-mansion/universal-sierra-compiler/blob/master/scripts/install.sh) binary
+
+#### Fixed
+
+- Error message for tests that should panic but pass
+
+## [0.14.0] - 2024-01-11
+
+### Forge
+
+#### Added
+
 - `Secp256k1` and `Secp256r1` curves support for `KeyPair` in `snforge_std`
 
 #### Changed